Transformation: move endpoint into api/rest package
