+++ transformation/src/index.ts
@@ -1,4 +1,4 @@
-import { TransformationEndpoint } from './transformationEndpoint'
+import { TransformationEndpoint } from './api/rest/transformationEndpoint'
 import VM2SandboxExecutor from './vm2SandboxExecutor'
 import JSTransformationService from './jsTransformationService'
